requirements.txt
django==4.0
environs==11.0.0
django-phonenumber-field[phonenumbers]
dj_database_url==2.1.0
psycopg2-binary==2.9.9